refactor: move Slice components to their own dir
