+++ src/slices/ImageGallery.js
