+++ src/slices/Quote.js
